+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     * Fixed crash caused by stale model indices
     * Precompositions correctly update whhen the current frame is changed
     * Clicking on the checkboxes in the pl8ugin settings properly toggles the plugins
+    * Lottie precomposition dependencies are correctly resolved on import
 
 ## 0.4.2